Naming hell hopefully solved
Problem when testing from<->to the same host detected (unrelated).

If allowed to run as is and in a screen, nc will bail and a second try will
show:

nc: Address already in use

This error sticks and -P must be used to choose another port. Suspect timing
issue, will try changing the order so that start is slightly faster.

Change-Id: Id3eee56ab4f8031723f80e7e9dc0a1b2885c1490
Change-Id: Ic01abb9adf80b4b6e335ceb888294846a547feea
Change-Id: Ie7e6c3b0473a3bcacb82aaf35210720c2269b1af